Add no_scale option to CLI arguments
This change adds a new boolean argument `no_scale` to the command-line interface (CLI) arguments. The purpose of this addition is to allow users to specify whether they want to disable image scaling when processing images. This can be useful in scenarios where preserving the aspect ratio is necessary, such as for web use or certain types of static content generation.